define mode as LowCardinality(String)
mode only takes on a handful of values; defining it as LowCardinality
uses less storage and also makes querying a bit more efficient.
Basically it's a dynamic enum, where strings are mapped to ints, and the
mapping is stored in another table (which is small enough to fit very
easily in memory).